# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/servicos": {
      "get": {
        "tags": [
          "Serviços"
        ],
        "summary": "Listar todos os serviços",
        "description": "Retorna a lista completa de serviços cadastrados",
        "operationId": "listar",
        "responses": {
          "200": {
            "description": "Lista de serviços retornada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/servicos",
          "httpMethod": "GET",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      },
      "post": {
        "tags": [
          "Serviços"
        ],
        "summary": "Cadastrar serviço",
        "description": "Cadastra um novo serviço no sistema",
        "operationId": "cadastrar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CadastrarServicoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Serviço cadastrado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/servicos",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
    "/pecas": {
      "get": {
        "tags": [
          "Peças e Insumos"
        ],
        "summary": "Listar todas as peças e insumos",
        "operationId": "listarTodas",
        "parameters": [
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0
            }
          },
          {
            "name": "quantidade",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de peças e insumos retornada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/pecas",
          "httpMethod": "GET",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      },
      "post": {
        "tags": [
          "Peças e Insumos"
        ],
        "summary": "Cadastrar nova peça ou insumo",
        "operationId": "cadastrar_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CadastrarPecaInsumoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Peça/insumo cadastrado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/pecas",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
...
    "/ordens-servicos/{id}/execucoes": {
      "post": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Inicia a execução de uma ordem de serviço após aprovação do orçamento",
        "operationId": "executar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutarOrdemServicoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Execução da OS iniciada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Ordem de serviço não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "Orçamento não aprovado ou OS em estado incorreto",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/ordens-servicos/{id}/execucoes",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
...
    "/clientes": {
      "post": {
        "tags": [
          "Clientes"
        ],
        "summary": "Cadastra um novo cliente",
        "operationId": "cadastrar_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CadastrarClienteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Cliente cadastrado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/clientes",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
...
    "/servicos/{id}": {
      "get": {
        "tags": [
          "Serviços"
        ],
        "summary": "Buscar serviço por ID",
        "description": "Retorna os detalhes de um serviço específico",
        "operationId": "buscar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do serviço",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Serviço encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Serviço não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/servicos/{id}",
          "httpMethod": "GET",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      },
      "delete": {
        "tags": [
          "Serviços"
        ],
        "summary": "Deletar serviço",
        "description": "Remove um serviço pelo seu ID",
        "operationId": "deletar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do serviço",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Serviço deletado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Serviço não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/servicos/{id}",
          "httpMethod": "DELETE",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      },
      "patch": {
        "tags": [
          "Serviços"
        ],
        "summary": "Atualizar serviço",
        "description": "Atualiza os dados de um serviço existente",
        "operationId": "atualizar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do serviço",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarServicoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Serviço atualizado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Serviço não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/servicos/{id}",
          "httpMethod": "PATCH",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
    "/pecas/{id}": {
      "get": {
        "tags": [
          "Peças e Insumos"
        ],
        "summary": "Buscar peça/insumo por ID",
        "operationId": "buscarPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Peça/insumo encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Peça/insumo não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/pecas/{id}",
          "httpMethod": "GET",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      },
      "delete": {
        "tags": [
          "Peças e Insumos"
        ],
        "summary": "Deletar peça ou insumo",
        "operationId": "deletar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Peça/insumo deletado com sucesso"
          },
          "404": {
            "description": "Peça/insumo não encontrado"
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/pecas/{id}",
          "httpMethod": "DELETE",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      },
      "patch": {
        "tags": [
          "Peças e Insumos"
        ],
        "summary": "Atualizar peça ou insumo",
        "operationId": "atualizar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarPecaInsumoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Peça/insumo atualizado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Peça/insumo não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/pecas/{id}",
          "httpMethod": "PATCH",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
    "/clientes/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Busca um cliente por ID",
        "operationId": "buscar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/clientes/{id}",
          "httpMethod": "GET",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      },
      "delete": {
        "tags": [
          "Clientes"
        ],
        "summary": "Deleta um cliente por ID",
        "operationId": "deletar_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cliente deletado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/clientes/{id}",
          "httpMethod": "DELETE",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      },
      "patch": {
        "tags": [
          "Clientes"
        ],
        "summary": "Atualiza os dados de um cliente",
        "operationId": "atualizar_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarClienteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Cliente atualizado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/clientes/{id}",
          "httpMethod": "PATCH",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CadastrarServicoRequest": {
        "type": "object",
        "description": "Dados para cadastro do serviço",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1
          },
          "descricao": {
            "type": "string",
            "minLength": 1
          },
          "preco": {
            "type": "number"
          },
          "duracaoEstimada": {
            "type": "integer",
            "format": "int32"
          },
          "categoria": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "categoria",
          "descricao",
          "duracaoEstimada",
          "nome",
          "preco"
        ]
      },
      "CadastrarPecaInsumoRequest": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1
          },
          "descricao": {
            "type": "string",
            "minLength": 1
          },
          "codigoFabricante": {
            "type": "string",
            "minLength": 1
          },
          "marca": {
            "type": "string",
            "minLength": 1
          },
          "quantidadeEstoque": {
            "type": "integer",
            "format": "int32"
          },
          "precoCusto": {
            "type": "number"
          },
          "precoVenda": {
            "type": "number"
          },
          "categoria": {
            "type": "string",
            "minLength": 1
          },
          "ativo": {
            "type": "boolean"
          }
        },
        "required": [
          "ativo",
          "categoria",
          "codigoFabricante",
          "descricao",
          "marca",
          "nome",
          "precoCusto",
          "precoVenda"
        ]
      },
...
      "ExecutarOrdemServicoRequest": {
        "type": "object",
        "properties": {
          "orcamentoId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CadastrarClienteRequest": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1
          },
          "sobrenome": {
            "type": "string",
            "minLength": 1
          },
          "cpfCnpj": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "telefone": {
            "type": "string",
            "minLength": 1
          },
          "endereco": {
            "$ref": "#/components/schemas/EnderecoRequest"
          }
        },
        "required": [
          "cpfCnpj",
          "email",
          "endereco",
          "nome",
          "sobrenome",
          "telefone"
        ]
      },
      "EnderecoRequest": {
        "type": "object",
        "properties": {
          "rua": {
            "type": "string",
            "minLength": 1
          },
          "numero": {
            "type": "string",
            "minLength": 1
          },
          "bairro": {
            "type": "string",
            "minLength": 1
          },
          "cidade": {
            "type": "string",
            "minLength": 1
          },
          "cep": {
            "type": "string",
            "minLength": 1
          },
          "uf": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "bairro",
          "cep",
          "cidade",
          "numero",
          "rua",
          "uf"
        ]
      },
...
      "AtualizarServicoRequest": {
        "type": "object",
        "description": "Dados para atualização do serviço",
        "properties": {
          "preco": {
            "type": "number"
          },
          "ativo": {
            "type": "boolean"
          }
        },
        "required": [
          "ativo",
          "preco"
        ]
      },
      "AtualizarPecaInsumoRequest": {
        "type": "object",
        "properties": {
          "qtd": {
            "type": "integer",
            "format": "int32"
          },
          "precoCusto": {
            "type": "number"
          },
          "precoVenda": {
            "type": "number"
          },
          "ativa": {
            "type": "boolean"
          }
        },
        "required": [
          "ativa",
          "precoCusto",
          "precoVenda",
          "qtd"
        ]
      },
      "AtualizarClienteRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "endereco": {
            "$ref": "#/components/schemas/EnderecoAtualizarRequest"
          }
        }
      },
      "EnderecoAtualizarRequest": {
        "type": "object",
        "properties": {
          "rua": {
            "type": "string",
            "minLength": 1
          },
          "numero": {
            "type": "string",
            "minLength": 1
          },
          "bairro": {
            "type": "string",
            "minLength": 1
          },
          "cep": {
            "type": "string",
            "minLength": 1
          },
          "cidade": {
            "type": "string",
            "minLength": 1
          },
          "uf": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "bairro",
          "cep",
          "cidade",
          "numero",
          "rua",
          "uf"
        ]
      },
...
    }
  }
}